requirements.txt
asgiref==3.5.2
Django==4.1.3
sqlparse==0.4.3
python-dotenv==0.20.0
psycopg2==2.9.3
gunicorn==20.1.0
django-filter==22.1
djangorestframework==3.13.1
drf-spectacular==0.22.1
django-cors-headers==3.10.0
django-silk==5.0.0
Pillow==8.4.0